refactor: move load_function to serializer module
