+++ src/autora/workflow/__main__.py
@@ -1,4 +1,3 @@
-import importlib
 import logging
 import pathlib
 from typing import Optional
@@ -7,7 +6,7 @@ import typer
 from typing_extensions import Annotated
 
 from autora.workflow.serializer import Supported as SerializersSupported
-from autora.workflow.serializer import dump_state, load_state
+from autora.workflow.serializer import _load_function, dump_state, load_state
 
 _logger = logging.getLogger(__name__)
 
@@ -58,15 +57,6 @@ def _configure_logger(debug, verbose):
     if verbose:
         logging.basicConfig(level=logging.INFO)
         _logger.info("using INFO logging level")
-
-
-def _load_function(fully_qualified_function_name: str):
-    _logger.debug(f"_load_function: Loading function {fully_qualified_function_name}")
-    module_name, function_name = fully_qualified_function_name.rsplit(".", 1)
-    module = importlib.import_module(module_name)
-    function = getattr(module, function_name)
-    _logger.debug(f"_load_function: Loaded function {function} from {module}")
-    return function
 
 
 if __name__ == "__main__":
